テスト関連コードを関数化
